Base class extracted as it is much likely that code gets re-used (e.g. for tasks)
